generate meal fragment

+++ data/src/main/java/com/example/data/models/generate_template/NutrientsTemplateResponse.kt
@@ -3,8 +3,8 @@ package com.example.data.models.generate_template
 import com.google.gson.annotations.SerializedName
 
 data class NutrientsTemplateResponse(
-    @SerializedName("calories") val calories: Int? = null,
-    @SerializedName("carbohydrates") val carbohydrates: Int? = null,
-    @SerializedName("fat") val fat: Int? = null,
-    @SerializedName("protein") val protein: Int? = null,
+    @SerializedName("calories") val calories: Double? = null,
+    @SerializedName("carbohydrates") val carbohydrates: Double? = null,
+    @SerializedName("fat") val fat: Double? = null,
+    @SerializedName("protein") val protein: Double? = null,
 )
