+++ domain/src/main/java/com/example/domain/models/generate_template/NutrientsTemplateData.kt
@@ -1,8 +1,8 @@
 package com.example.domain.models.generate_template
 
 data class NutrientsTemplateData(
-    val calories: Int,
-    val carbohydrates: Int,
-    val fat: Int,
-    val protein: Int
+    val calories: Double,
+    val carbohydrates: Double,
+    val fat: Double,
+    val protein: Double
 )
